# api-docs/openapi-device-base-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "设备基础数据业务模块 API 文档",
    "description": "weili-iot-portal 业务模块中设备基础数据（device-base）模块的 REST API 接口文档。\n\n功能覆盖设备基础信息、设备配置、设备型号、设备类型、组织单元以及设备关系等主数据维护。\n",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "/api/device-base/v1",
      "description": "设备基础数据 API 基础路径"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/device-types/{id}": {
      "get": {
        "tags": [
          "设备类型"
        ],
        "summary": "查询设备类型详情",
        "operationId": "getDeviceType",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "设备类型ID"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultDeviceTypeVO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "设备类型"
        ],
        "summary": "更新设备类型",
        "operationId": "updateDeviceType",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "设备类型ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceTypeUpdateReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultDeviceTypeVO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "设备类型"
        ],
        "summary": "删除设备类型",
        "operationId": "deleteDeviceType",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "设备类型ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/device-types/page": {
      "post": {
        "tags": [
          "设备类型"
        ],
        "summary": "分页查询设备类型",
        "operationId": "pageDeviceType",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceTypeQueryReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultPageDeviceTypeVO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/device-relations/{id}": {
      "get": {
        "tags": [
          "设备关系"
        ],
        "summary": "查询设备关系详情",
        "operationId": "getDeviceRelation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "关系ID"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultDeviceRelationVO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "设备关系"
        ],
        "summary": "更新设备关系",
        "operationId": "updateDeviceRelation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "关系ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRelationUpdateReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultDeviceRelationVO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "设备关系"
        ],
        "summary": "删除设备关系",
        "operationId": "deleteDeviceRelation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "关系ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/device-relations/page": {
      "post": {
        "tags": [
          "设备关系"
        ],
        "summary": "分页查询设备关系",
        "operationId": "pageDeviceRelation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRelationQueryReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultPageDeviceRelationVO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CommonResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码（200 表示成功）"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "时间戳"
          }
        }
      },
      "CommonResultBoolean": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonResult"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "boolean"
              }
            }
          }
        ]
      },
      "PageResult": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "pageNo": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "records": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "DeviceTypeVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "设备类型ID（主键）"
          },
          "typeCode": {
            "type": "string",
            "description": "类型编码（唯一标识）"
          },
          "typeName": {
            "type": "string",
            "description": "类型名称（显示名称）"
          },
          "parentTypeId": {
            "type": "string",
            "description": "父类型ID（用于树形结构，根节点为null）"
          },
          "parentTypeName": {
            "type": "string",
            "description": "父类型名称（冗余字段，便于展示）"
          },
          "level": {
            "type": "integer",
            "description": "层级（1-主类型，2-子类型）"
          },
          "category": {
            "type": "string",
            "description": "分类（如：CNC、PLC等）"
          },
          "description": {
            "type": "string",
            "description": "类型描述（详细说明）"
          },
          "icon": {
            "type": "string",
            "description": "图标（图标URL或图标代码）"
          },
          "customFields": {
            "type": "object",
            "additionalProperties": true,
            "description": "自定义字段（JSON对象，存储扩展属性）"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用（true-启用，false-停用）"
          },
          "sortOrder": {
            "type": "integer",
            "description": "排序顺序（数字越小越靠前）"
          },
          "createdTime": {
            "type": "integer",
            "format": "int64",
            "description": "创建时间（毫秒时间戳）"
          },
          "updatedTime": {
            "type": "integer",
            "format": "int64",
            "description": "更新时间（毫秒时间戳）"
          }
        }
      },
...
      "DeviceTypeUpdateReq": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "设备类型ID（必填，用于更新）"
          },
          "typeName": {
            "type": "string",
            "description": "类型名称（可选）"
          },
          "description": {
            "type": "string",
            "description": "类型描述（可选）"
          },
          "icon": {
            "type": "string",
            "description": "图标（可选）"
          },
          "customFields": {
            "type": "object",
            "additionalProperties": true,
            "description": "自定义字段（可选，JSON对象）"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用（可选）"
          },
          "sortOrder": {
            "type": "integer",
            "description": "排序顺序（可选）"
          }
        }
      },
      "DeviceTypeQueryReq": {
        "type": "object",
        "required": [
          "pageNo",
          "pageSize"
        ],
        "properties": {
          "typeCodeLike": {
            "type": "string",
            "description": "类型编码模糊查询（可选）"
          },
          "typeNameLike": {
            "type": "string",
            "description": "类型名称模糊查询（可选）"
          },
          "parentTypeId": {
            "type": "string",
            "description": "父类型ID（可选，查询指定父类型下的子类型）"
          },
          "level": {
            "type": "integer",
            "description": "层级查询（可选，1-主类型，2-子类型）"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "分类列表（可选，支持多选）"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用（可选，true-仅查询启用，false-仅查询停用，不传-不过滤）"
          },
          "pageNo": {
            "type": "integer",
            "minimum": 1,
            "description": "页码（必填，从1开始）"
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200,
            "description": "每页条数（必填，1-200）"
          },
          "sortBy": {
            "type": "string",
            "description": "排序字段（可选）"
          },
          "sortDirection": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ],
            "description": "排序方向（可选，ASC-升序，DESC-降序）"
          }
        }
      },
      "CommonResultDeviceTypeVO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonResult"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/DeviceTypeVO"
              }
            }
          }
        ]
      },
      "CommonResultPageDeviceTypeVO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonResult"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/PageResult"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "records": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/DeviceTypeVO"
                        }
                      }
                    }
                  }
                ]
              }
            }
          }
        ]
      },
...
      "DeviceRelationVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "设备关系ID（主键）"
          },
          "fromDeviceId": {
            "type": "string",
            "description": "源设备ID（关系起始设备）"
          },
          "fromDeviceCode": {
            "type": "string",
            "description": "源设备编号（冗余字段，便于展示）"
          },
          "fromDeviceName": {
            "type": "string",
            "description": "源设备名称（冗余字段，便于展示）"
          },
          "toDeviceId": {
            "type": "string",
            "description": "目标设备ID（关系目标设备）"
          },
          "toDeviceCode": {
            "type": "string",
            "description": "目标设备编号（冗余字段，便于展示）"
          },
          "toDeviceName": {
            "type": "string",
            "description": "目标设备名称（冗余字段，便于展示）"
          },
          "relationType": {
            "type": "string",
            "description": "关系类型（如：PARENT_CHILD-父子关系、PEER-同级关系、DEPENDENCY-依赖关系等）"
          },
          "relationName": {
            "type": "string",
            "description": "关系名称（关系的显示名称）"
          },
          "description": {
            "type": "string",
            "description": "关系描述（关系的详细说明）"
          },
          "properties": {
            "type": "object",
            "additionalProperties": true,
            "description": "关系属性（JSON对象，存储关系的扩展属性）"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用（true-启用，false-停用）"
          },
          "createdTime": {
            "type": "integer",
            "format": "int64",
            "description": "创建时间（毫秒时间戳）"
          },
          "updatedTime": {
            "type": "integer",
            "format": "int64",
            "description": "更新时间（毫秒时间戳）"
          },
          "createdBy": {
            "type": "string",
            "description": "创建人（用户ID或用户名）"
          },
          "updatedBy": {
            "type": "string",
            "description": "更新人（用户ID或用户名）"
          }
        }
      },
...
      "DeviceRelationUpdateReq": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "设备关系ID（必填，用于更新）"
          },
          "relationName": {
            "type": "string",
            "description": "关系名称（可选）"
          },
          "description": {
            "type": "string",
            "description": "关系描述（可选）"
          },
          "properties": {
            "type": "object",
            "additionalProperties": true,
            "description": "关系属性（可选，JSON对象）"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用（可选）"
          }
        }
      },
      "DeviceRelationQueryReq": {
        "type": "object",
        "required": [
          "pageNo",
          "pageSize"
        ],
        "properties": {
          "fromDeviceId": {
            "type": "string",
            "description": "源设备ID（可选，查询指定设备作为源的关系）"
          },
          "toDeviceId": {
            "type": "string",
            "description": "目标设备ID（可选，查询指定设备作为目标的关系）"
          },
          "relationTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "关系类型列表（可选，支持多选）"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用（可选，true-仅查询启用，false-仅查询停用，不传-不过滤）"
          },
          "pageNo": {
            "type": "integer",
            "minimum": 1,
            "description": "页码（必填，从1开始）"
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200,
            "description": "每页条数（必填，1-200）"
          },
          "sortBy": {
            "type": "string",
            "description": "排序字段（可选）"
          },
          "sortDirection": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ],
            "description": "排序方向（可选，ASC-升序，DESC-降序）"
          }
        }
      },
      "CommonResultDeviceRelationVO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonResult"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/DeviceRelationVO"
              }
            }
          }
        ]
      },
      "CommonResultPageDeviceRelationVO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonResult"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/PageResult"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "records": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/DeviceRelationVO"
                        }
                      }
                    }
                  }
                ]
              }
            }
          }
        ]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CommonResult"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未授权访问",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CommonResult"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CommonResult"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "服务器内部错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CommonResult"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer Token 认证"
      }
    }
  }
}